Move DataStore to data_store
